+++ include/pollard.h
@@ -1,6 +1,6 @@
 #ifndef POLLARD_H
 #define POLLARD_H
 
-int pollard(int n);
+long long pollard(int n);
 
 #endif
